Windows build fixes and removal of a bunch of warnings

+++ src/aax_effects.c
@@ -618,7 +618,7 @@ aaxEffectSetState(aaxEffect e, int state)
                         } else if (data->lfo.value[t] > data->lfo.max) {
                            data->lfo.value[t] = data->lfo.max;
                         }
-                        data->delay.sample_offs[t] = data->lfo.value[t];
+                        data->delay.sample_offs[t] = (unsigned int)data->lfo.value[t];
 
                         switch (state & ~AAX_INVERSE)
                         {
@@ -669,7 +669,7 @@ aaxEffectSetState(aaxEffect e, int state)
                      for (t=0; t<_AAX_MAX_SPEAKERS; t++)
                      {
                         data->lfo.value[t] = data->lfo.min;
-                        data->delay.sample_offs[t] = data->lfo.value[t];
+                        data->delay.sample_offs[t] = (unsigned int)data->lfo.value[t];
                      }
                      data->lfo.get = _oalRingBufferLFOGetFixedValue;
                   }
