+++ src/software/rbuf_mixers.c
@@ -381,7 +381,7 @@ bufCompress(void *d, unsigned int dmin, unsigned int dmax, float clip, float asy
       int32_t samp;
 
       samp = *ptr;
-      asamp = fabsf(samp+asym);
+      asamp = (int32_t)fabsf(samp+asym);
 
       pos = 1+(asamp >> SHIFT);
       sdf = _MINMAX(asamp*df, 0.0f, 1.0f);
